Refactor Pressure BCs (#30178)
